add iterator benchmark
