+++ storage/operation/writes_bench_test.go
@@ -10,20 +10,6 @@ import (
 	"github.com/onflow/flow-go/storage/operation/dbtest"
 )
 
-func BenchmarkRetrieve(t *testing.B) {
-	dbtest.BenchWithStorages(t, func(t *testing.B, r storage.Reader, withWriter dbtest.WithWriter) {
-		e := Entity{ID: 1337}
-		require.NoError(t, withWriter(operation.Upsert(e.Key(), e)))
-
-		t.ResetTimer()
-
-		for i := 0; i < t.N; i++ {
-			var readBack Entity
-			require.NoError(t, operation.Retrieve(e.Key(), &readBack)(r))
-		}
-	})
-}
-
 func BenchmarkUpsert(t *testing.B) {
 	dbtest.BenchWithStorages(t, func(t *testing.B, r storage.Reader, withWriter dbtest.WithWriter) {
 		for i := 0; i < t.N; i++ {
